projects/ASL-fingerspelling/Detection/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
flask==2.0.3
werkzeug==2.0.3
flask-cors==3.0.10
numpy==1.24.2
torch>=1.13.1
transformers>=4.26.0
pillow>=9.4.0
opencv-python==4.7.0.72
matplotlib